fix specs + add torch refs

+++ challenges/elu/template.cu
@@ -1,6 +1,6 @@
 #include <cuda_runtime.h>
 
 // ELU.
-void solve(const float *input, float *output, int n, int m, int alpha) {
+void solve(const float *input, float *output, int n, int m, float alpha) {
 
 }
