+++ challenges/threshold/template.cu
@@ -1,6 +1,6 @@
 #include <cuda_runtime.h>
 
 // Image Thresholding.
-void solve(const float *input_image, int threshold_value, float *output_image, int height, int width) {
+void solve(const float *input_image, float threshold_value, float *output_image, int height, int width) {
 
 }
